# Conflicts:
#	.idea/workspace.xml
#	out/artifacts/ProjectAutotest_jar/ProjectAutotest.jar
#	src/main/java/screen/AT2MDMSY1002/AT2MDMSY1002Sis.java
#	target/classes/core/CommonActions/Functions.class

+++ src/main/java/screen/AT2ACCSU0075/AT2ACCSU0075Test.java
@@ -14,38 +14,30 @@ public class AT2ACCSU0075Test {
 
     public AT2ACCSU0075Test() {
     }
-
     public AT2ACCSU0075Locators getLocators() {
         return locators;
     }
-
     public void setLocators(AT2ACCSU0075Locators locators) {
         this.locators = locators;
     }
-
     public AT2ACCSU0075Data getData() {
         return data;
     }
-
     public void setData(AT2ACCSU0075Data data) {
         this.data = data;
     }
-
     public void start(TestDriver driver) {
         setScreenInfo(driver);
         CommonProcedures.goToScreen(driver);
     }
-
     protected void setScreenInfo(TestDriver driver) {
-        driver.getTestdetails().setMainmenu("Setup");
+        driver.getTestdetails().setMainmenu("Accommodation");
         driver.getTestdetails().setSubmenu("Setup");
         driver.getTestdetails().setScreen("Atributos plantillas");
     }
-
     protected String getElements(String key) {
         return String.valueOf(this.locators.getElements().get(key));
     }
-
     protected String getData(String key) {
         return String.valueOf(this.data.getData().get(key));
     }
@@ -54,6 +46,7 @@ public class AT2ACCSU0075Test {
         if (!interaction_record_accommodation(driver)) return false;
         if (!search_accommodation(driver)) return false;
         if (!interaction_edit_accommodation(driver)) return false;
+        if (!search_reset(driver)) return false;
         if (!qbe_accommodation(driver)) return false;
         if (!others_actions_accommodation(driver)) return false;
         if (!delete_accommodation(driver)) return false;
@@ -78,22 +71,22 @@ public class AT2ACCSU0075Test {
             return false;
         }
         if (!Functions.createLov(driver,
-                new String[]{"add_lov_concepto", getElements("add_lov_concepto")}, // b_lov
-                new String[]{"add_i_concepto", getElements("add_i_concepto")}, // i_lov
-                recursiveXPaths.lov_b_search, // lov b search
-                recursiveXPaths.lov_e_result, // lov result
-                recursiveXPaths.lov_b_ok, //lov b ok
-                "concepto", //Data name
-                where)) {
-            return false;
-        }
-        if (!Functions.createLov(driver,
                 new String[]{"add_lov_grupo", getElements("add_lov_grupo")}, // b_lov
                 new String[]{"add_i_grupo", getElements("add_i_grupo")}, // i_lov
                 recursiveXPaths.lov_b_search, // lov b search
                 recursiveXPaths.lov_e_result, // lov result
                 recursiveXPaths.lov_b_ok, //lov b ok
                 "grupo", //Data name
+                where)) {
+            return false;
+        }
+        if (!Functions.createLov(driver,
+                new String[]{"add_lov_concepto", getElements("add_lov_concepto")}, // b_lov
+                new String[]{"add_i_concepto", getElements("add_i_concepto")}, // i_lov
+                recursiveXPaths.lov_b_search, // lov b search
+                recursiveXPaths.lov_e_result, // lov result
+                recursiveXPaths.lov_b_ok, //lov b ok
+                "concepto", //Data name
                 where)) {
             return false;
         }
@@ -105,7 +98,6 @@ public class AT2ACCSU0075Test {
         }
         return true;
     }
-
     private boolean search_accommodation(TestDriver driver) {
         driver.getReport().addHeader("SEARCH RECORD", 3, false);
         Functions.break_time(driver, 30, 500);
@@ -119,15 +111,6 @@ public class AT2ACCSU0075Test {
             return false;
         }
         if (!Functions.createLovByValue(driver,
-                new String[]{"search_lov_concepto", getElements("search_lov_concepto")}, //LoV button
-                new String[]{"search_i_concepto", getElements("search_i_concepto")}, //external LoV input
-                new String[]{"search_lov_concepto_i_code", recursiveXPaths.lov_i_genericinput}, //internal LoV input
-                getData("concepto"), // value to search
-                "concepto", //name of the data
-                where)) {
-            return false;
-        }
-        if (!Functions.createLovByValue(driver,
                 new String[]{"search_lov_grupo", getElements("search_lov_grupo")}, //LoV button
                 new String[]{"search_i_grupo", getElements("search_i_grupo")}, //external LoV input
                 new String[]{"search_lov_grupo_i_code", recursiveXPaths.lov_i_genericinput}, //internal LoV input
@@ -136,6 +119,16 @@ public class AT2ACCSU0075Test {
                 where)) {
             return false;
         }
+        if (!Functions.createLovByValue(driver,
+                new String[]{"search_lov_concepto", getElements("search_lov_concepto")}, //LoV button
+                new String[]{"search_i_concepto", getElements("search_i_concepto")}, //external LoV input
+                new String[]{"search_lov_concepto_i_code", recursiveXPaths.lov_i_genericinput}, //internal LoV input
+                getData("concepto"), // value to search
+                "concepto", //name of the data
+                where)) {
+            return false;
+        }
+
         if (!Functions.clickSearchAndResult(driver,
                 new String[]{"search_b_search", getElements("search_b_search")}, //search button
                 new String[]{"setup_e_result", getElements("setup_e_result")}, //result element
@@ -144,15 +137,9 @@ public class AT2ACCSU0075Test {
         }
         return true;
     }
-
     private boolean interaction_edit_accommodation(TestDriver driver) {
         driver.getReport().addHeader("EDITION RECORD", 3, false);
         String where = " on EDITION";
-        if (!Functions.simpleClick(driver,
-                new String[]{"search_b_reset", getElements("search_b_reset")}, //element to click
-                where)) {
-            return false;
-        }
         if (!Functions.checkClick(driver,
                 new String[]{"setup_b_edit", getElements("setup_b_edit")}, //element to click
                 recursiveXPaths.glass, //element expected to appear
@@ -168,22 +155,22 @@ public class AT2ACCSU0075Test {
             return false;
         }
         if (!Functions.createLov(driver,
-                new String[]{"add_lov_concepto", getElements("add_lov_concepto")}, // b_lov
-                new String[]{"add_i_concepto", getElements("add_i_concepto")}, // i_lov
-                recursiveXPaths.lov_b_search, // lov b search
-                recursiveXPaths.lov_e_altresult, // lov result
-                recursiveXPaths.lov_b_ok, //lov b ok
-                "concepto", //Data name
-                where)) {
-            return false;
-        }
-        if (!Functions.createLov(driver,
                 new String[]{"add_lov_grupo", getElements("add_lov_grupo")}, // b_lov
                 new String[]{"add_i_grupo", getElements("add_i_grupo")}, // i_lov
                 recursiveXPaths.lov_b_search, // lov b search
                 recursiveXPaths.lov_e_altresult, // lov result
                 recursiveXPaths.lov_b_ok, //lov b ok
                 "grupo", //Data name
+                where)) {
+            return false;
+        }
+        if (!Functions.createLov(driver,
+                new String[]{"add_lov_concepto", getElements("add_lov_concepto")}, // b_lov
+                new String[]{"add_i_concepto", getElements("add_i_concepto")}, // i_lov
+                recursiveXPaths.lov_b_search, // lov b search
+                recursiveXPaths.lov_e_altresult, // lov result
+                recursiveXPaths.lov_b_ok, //lov b ok
+                "concepto", //Data name
                 where)) {
             return false;
         }
@@ -195,7 +182,16 @@ public class AT2ACCSU0075Test {
         }
         return true;
     }
-
+    private boolean search_reset(TestDriver driver) {
+        driver.getReport().addHeader("SEARCH RESET", 3, false);
+        String where = " on SEARCH RESET";
+        if (!Functions.simpleClick(driver,
+                new String[]{"search_b_reset", getElements("search_b_reset")}, //element to click
+                where)) {
+            return false;
+        }
+        return true;
+    }
     private boolean qbe_accommodation(TestDriver driver) {
         driver.getReport().addHeader("QBE RECORD", 3, false);
         String where = " on QBE";
@@ -213,12 +209,12 @@ public class AT2ACCSU0075Test {
                 "negocio", getData("negocio"), where)) {
             return false;
         }
-        if (!Functions.insertInput(driver, new String[]{"qbe_i_concepto", getElements("qbe_i_concepto")}, // element path
-                "concepto", getData("concepto"), where)) {
-            return false;
-        }
         if (!Functions.insertInput(driver, new String[]{"qbe_i_grupo", getElements("qbe_i_grupo")}, // element path
                 "grupo", getData("grupo"), where)) {
+            return false;
+        }
+        if (!Functions.insertInput(driver, new String[]{"qbe_i_concepto", getElements("qbe_i_concepto")}, // element path
+                "concepto", getData("concepto"), where)) {
             return false;
         }
         if (!Functions.enterQueryAndClickResult(driver,
@@ -229,7 +225,6 @@ public class AT2ACCSU0075Test {
         }
         return true;
     }
-
     private boolean others_actions_accommodation(TestDriver driver) {
         driver.getReport().addHeader("OTHER ACTIONS - DETACH", 3, false);
         String where = " on DETACH";
@@ -241,7 +236,6 @@ public class AT2ACCSU0075Test {
         }
         return true;
     }
-
     private boolean delete_accommodation(TestDriver driver) {
         driver.getReport().addHeader("DELETE DATA", 3, false);
         String where = " on DELETE DATA";
